+++ viadot/tasks/__init__.py
@@ -40,5 +40,5 @@ except ImportError:
     pass
 
 from .duckdb import DuckDBCreateTableFromParquet, DuckDBQuery, DuckDBToDF
-from .sql_server import SQLServerCreateTable, SQLServerToDF
+from .sql_server import SQLServerCreateTable, SQLServerToDF, SQLServerQuery
 from .epicor import EpicorOrdersToDF